expanded on repo initialization
Next work in main.go but essentially.
Find .env files in each request  folders add this as TODO.
initialize the project.
be able to set the default env to the os.SetEnv() with a flag of -env otherwise default is global.
Also sort out which one wins Global < collection < default env
Be able to read the values from the env files.
Tests